+++ src/QRCodeCapturer.h
