+++ src/SSValidator.h
